+++ backend/canisters/local_user_index/impl/src/guards.rs
@@ -72,14 +72,6 @@ pub fn caller_is_notification_pusher() -> Result<(), String> {
     }
 }
 
-pub fn caller_is_notifications_canister() -> Result<(), String> {
-    if read_state(|state| state.is_caller_notifications_canister()) {
-        Ok(())
-    } else {
-        Err("Caller is not a notifications canister".to_string())
-    }
-}
-
 pub fn caller_is_openchat_user() -> Result<(), String> {
     if read_state(|state| state.is_caller_openchat_user()) {
         Ok(())