+++ backend/canisters/notifications/api/src/queries/mod.rs
@@ -1,4 +1,0 @@
-pub mod latest_notification_index;
-pub mod notification_types;
-pub mod notifications;
-pub mod notifications_v2;