+++ backend/canisters/notifications/impl/src/queries/mod.rs
@@ -1,4 +1,0 @@
-mod http_request;
-mod latest_notification_index;
-mod notifications;
-mod notifications_v2;