+++ backend/canisters/notifications_index/api/src/updates/notify_local_index_added.rs
@@ -1,8 +1,10 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::BuildVersion;
+use types::{CanisterId, UnitResult};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub wasm_version: BuildVersion,
+    pub canister_id: CanisterId,
 }
+
+pub type Response = UnitResult;