+++ backend/canisters/notifications_index/c2c_client/src/lib.rs
@@ -4,5 +4,4 @@ use notifications_index_canister::*;
 // Queries
 
 // Updates
-generate_c2c_call!(add_notifications_canister);
-generate_c2c_call!(c2c_user_index, 300);
+generate_c2c_call!(notify_local_index_added);