+++ backend/canisters/notifications_index/client/src/lib.rs
@@ -4,7 +4,6 @@ use notifications_index_canister::*;
 // Queries
 
 // Updates
-generate_candid_update_call!(add_notifications_canister);
+generate_candid_update_call!(notify_local_index_added);
 generate_candid_update_call!(push_subscription);
 generate_candid_update_call!(remove_subscriptions);
-generate_candid_update_call!(upgrade_notifications_canister_wasm);