+++ backend/canisters/notifications_index/impl/src/updates/mod.rs
@@ -1,8 +1,6 @@
-mod add_notifications_canister;
-mod c2c_user_index;
+mod notify_local_index_added;
 mod push_subscription;
 mod remove_subscription;
 mod remove_subscriptions;
 mod remove_subscriptions_for_user;
-mod upgrade_notifications_canister_wasm;
 mod wallet_receive;