+++ backend/canisters/user_index/api/src/updates/add_local_user_index_canister.rs
@@ -6,7 +6,6 @@ use types::CanisterId;
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub canister_id: CanisterId,
-    pub notifications_canister_id: CanisterId,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]