+++ backend/canisters/user_index/impl/src/updates/c2c_local_user_index.rs
@@ -154,18 +154,10 @@ fn handle_event<F: FnOnce() -> TimestampMillis>(
         LocalUserIndexEvent::UserBlocked(user_id, blocked) => {
             state.data.blocked_users.insert((user_id, blocked), ());
             state.push_event_to_all_local_user_indexes(UserIndexEvent::UserBlocked(user_id, blocked), Some(caller));
-            state.push_event_to_notifications_index(
-                notifications_index_canister::UserIndexEvent::UserBlocked(user_id, blocked),
-                **now,
-            );
         }
         LocalUserIndexEvent::UserUnblocked(user_id, unblocked) => {
             state.data.blocked_users.remove(&(user_id, unblocked));
             state.push_event_to_all_local_user_indexes(UserIndexEvent::UserUnblocked(user_id, unblocked), Some(caller));
-            state.push_event_to_notifications_index(
-                notifications_index_canister::UserIndexEvent::UserUnblocked(user_id, unblocked),
-                **now,
-            );
         }
         LocalUserIndexEvent::SetMaxStreak(user_id, max_streak) => state.data.users.set_max_streak(&user_id, max_streak),
     }